pyproject.toml
[tool.poetry]
name = "items-manager"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "~3.7"
chalice = "^1.12"
boto3 = "^1.10.46"
marshmallow = "^3.3.0"

[tool.poetry.dev-dependencies]
"aws-cdk.core" = "^1.19"
"aws-cdk.aws-dynamodb" = "^1.19"
"aws-cdk.aws-iam" = "^1.19"
cdk-chalice = "^0.4.0"
pycodestyle = "^2.5.0"
mypy = "^0.761"
pylint = "^2.4.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
